Move Sidebar and OverlayPage
