+++ pages/[id]/index.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 import cx from 'classnames'
-import { Sidebar } from 'components/Sidebar'
+import { Sidebar } from 'components/editor/Sidebar'
 import { useUser } from '@auth0/nextjs-auth0'
 import { useFormFetch } from 'lib/hooks/useFormFetch'
 import { useRouter } from 'next/router'
@@ -15,7 +15,7 @@ import { useResponses } from 'lib/hooks/useResponses'
 import { LabelSwitch } from 'components/editor/LabelSwitch'
 import { mutate } from 'swr'
 import toast from 'react-hot-toast'
-import { OverlayPage } from 'components/OverlayPage'
+import { OverlayPage } from 'components/common/OverlayPage'
 import { Layout } from 'components/Layout'
 
 const FormDashboard: NextPage = () => {
